chore: add capsule_create_error_01 comments
